backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
pinecone-client==2.2.4
supabase==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.1